Category NSString+OPFEscapeStrings available
Use this category to escape string by predefined escaping schemes:

1. Import "NSString + OPFEscapeStrings.h" in your source file.
2. Call id escapedString = [yourString escapeWithScheme:OPFEscapeScheme]

OPFEscapeScheme can be OPFEscapePrettify for instance

If you want to add new escape Schemes:
OPFEscapeScheme is an enum defining the scheme of escaping. It is
defined in the category's .h file. The replacement arrays are defined
in the switch block in the category's .m file